refactor: separate category scoring responsibilities
